third_party/WebKit/LayoutTests/external/wpt/html/browsers/history/the-location-interface/per-global.window-expected.txt

This is a testharness.js-based test.
PASS Discarding the browsing context must not change window.location 
FAIL Navigating from the initial about:blank must not replace window.location assert_equals: expected [stringifying object threw TypeError: Cannot convert object to primitive value with type object] but got object "http://web-platform.test:8001/common/blank.html"
FAIL document.open() must replace window.location assert_not_equals: got disallowed value object "http://web-platform.test:8001/html/browsers/history/the-location-interface/per-global.window.html"
Harness: the test ran to completion.
